requirements.txt
flask~=3.0.2
flask-restful
flask_cors
pdf2image
torch~=2.2.0
protobuf==3.20.2
torchvision~=0.17.0
opencv-python~=4.6.0.66
scikit-image
scipy~=1.10.1
gitpython
matplotlib~=3.7.5
numpy~=1.24.4
pillow~=9.5.0
psutil  # system resources
PyYAML
requests~=2.31.0
thop  # FLOPs computation
tqdm~=4.66.2
faiss-cpu
six
pandas~=2.0.3
vietocr~=0.3.12
paddleocr~=2.7.0.3
paddlepaddle
ultralytics~=8.1.19
gradio~=4.19.2
pymongo~=4.6.2
python-dotenv~=1.0.1
uvicorn~=0.27.1
fastapi~=0.110.0
packaging~=23.2
seaborn~=0.13.2
openpyxl~=3.1.2